move Backup*Error to errors module
